requirements.txt
# Core Agent Framework
langgraph>=0.2.0
langchain-core
langchain-openai
langchain-anthropic

# Configuration & Utilities
python-dotenv
pathlib
dataclasses

# Tools & Shell
# (Most are standard library)

# REPL history (Windows only — macOS/Linux use the built-in readline)
pyreadline3; sys_platform == "win32"
